ffu: tripleo-packages repo management
This change introduces some basic repo management for fast-forward
upgrades using the tripleo-repos tool as the default implementation.

The following parameters have been added to the template to allow for
additional implementations to be added.

FastForwardRepoType - Currently defaults to tripleo-repos
FastForwardRepoArgs - Currently defaults to tripleo-repos args for O and P

bp fast-forward-upgrades
Change-Id: I92f6f5015f34e6c5e8ef131f303d9c8144d1c83e